Seconds cross to samples by one rule, so boxes that meet keep meeting
Placing the multitrack in seconds crossed a box's length on its own (dur / rate)
and its end as a sum, which differ in the last bit for most sample counts. A box
stopped ending exactly where the next began, so a snap sometimes missed and a
join, comparing seconds against f64::EPSILON, saw a gap and refused.

The rule is now one and lives in the core: a position lands on a whole sample
(secs_to_samples) and a length is the difference of its two ends, each crossed
(secs_to_samples_over / samples_to_secs_over). The multitrack projection and the
GUI host's look both go through it, and read_join asks where boxes meet in
samples at the view's rate.

The GUI host reported the multitrack's clip and curve positions as f32, which
holds a whole sample only up to 2^24 (under six minutes at 48 kHz); they are
doubles now.

A box drawn a different length than its samples, reported with the same message,
is not reproduced yet and is filed open in the document plan.

+++ clients/gui/src/host/elements/multitrack/props.rs
@@ -15,7 +15,10 @@ pub(super) fn json_arg(v: Value) -> OscType {
     match v {
         Value::String(s) => OscType::String(s),
         Value::Number(n) if n.is_i64() => OscType::Int(n.as_i64().unwrap_or(0) as i32),
-        other => OscType::Float(other.as_f64().unwrap_or(0.0) as f32),
+        // **A double, not a float**: these are positions in samples, and an
+        // `f32` holds a whole sample only up to 2^24 of them -- under six
+        // minutes at 48 kHz -- so a box past that came back somewhere else.
+        other => OscType::Double(other.as_f64().unwrap_or(0.0)),
     }
 }
